migrate core to new structure
