+++ app/lib/core/v2_0/doc-max.json
@@ -1,0 +1,334 @@
+{
+  "document": {
+    "acknowledgments": [
+      {
+        "names": [""],
+        "organization": "",
+        "summary": "",
+        "urls": [""]
+      }
+    ],
+    "aggregate_severity": {
+      "namespace": "",
+      "text": ""
+    },
+    "category": "",
+    "csaf_version": "2.0",
+    "distribution": {
+      "text": "",
+      "tlp": {
+        "label": "",
+        "url": "https://www.first.org/tlp/"
+      }
+    },
+    "lang": "",
+    "notes": [
+      {
+        "audience": "",
+        "category": "",
+        "text": "",
+        "title": ""
+      }
+    ],
+    "publisher": {
+      "category": "",
+      "contact_details": "",
+      "issuing_authority": "",
+      "name": "",
+      "namespace": ""
+    },
+    "references": [
+      {
+        "category": "external",
+        "summary": "",
+        "url": ""
+      }
+    ],
+    "source_lang": "",
+    "title": "",
+    "tracking": {
+      "aliases": [""],
+      "current_release_date": "",
+      "generator": {
+        "date": "",
+        "engine": {
+          "name": "",
+          "version": ""
+        }
+      },
+      "id": "",
+      "initial_release_date": "",
+      "revision_history": [
+        {
+          "date": "",
+          "legacy_version": "",
+          "number": "",
+          "summary": ""
+        }
+      ],
+      "status": "",
+      "version": ""
+    }
+  },
+  "product_tree": {
+    "branches": [
+      {
+        "branches": [],
+        "category": "",
+        "name": "",
+        "product": {
+          "name": "",
+          "product_id": "",
+          "product_identification_helper": {
+            "cpe": "",
+            "hashes": [
+              {
+                "file_hashes": [
+                  {
+                    "algorithm": "sha256",
+                    "value": ""
+                  }
+                ],
+                "filename": ""
+              }
+            ],
+            "model_numbers": [""],
+            "purl": "",
+            "sbom_urls": [""],
+            "serial_numbers": [""],
+            "skus": [""],
+            "x_generic_uris": [
+              {
+                "namespace": "",
+                "uri": ""
+              }
+            ]
+          }
+        }
+      }
+    ],
+    "full_product_names": [
+      {
+        "name": "",
+        "product_id": "",
+        "product_identification_helper": {
+          "cpe": "",
+          "hashes": [
+            {
+              "file_hashes": [
+                {
+                  "algorithm": "sha256",
+                  "value": ""
+                }
+              ],
+              "filename": ""
+            }
+          ],
+          "model_numbers": [""],
+          "purl": "",
+          "sbom_urls": [""],
+          "serial_numbers": [""],
+          "skus": [""],
+          "x_generic_uris": [
+            {
+              "namespace": "",
+              "uri": ""
+            }
+          ]
+        }
+      }
+    ],
+    "product_groups": [
+      {
+        "group_id": "",
+        "product_ids": ["", ""],
+        "summary": ""
+      }
+    ],
+    "relationships": [
+      {
+        "category": "",
+        "full_product_name": {
+          "name": "",
+          "product_id": "",
+          "product_identification_helper": {
+            "cpe": "",
+            "hashes": [
+              {
+                "file_hashes": [
+                  {
+                    "algorithm": "sha256",
+                    "value": ""
+                  }
+                ],
+                "filename": ""
+              }
+            ],
+            "model_numbers": [""],
+            "purl": "",
+            "sbom_urls": [""],
+            "serial_numbers": [""],
+            "skus": [""],
+            "x_generic_uris": [
+              {
+                "namespace": "",
+                "uri": ""
+              }
+            ]
+          }
+        },
+        "product_reference": "",
+        "relates_to_product_reference": ""
+      }
+    ]
+  },
+  "vulnerabilities": [
+    {
+      "acknowledgments": [
+        {
+          "names": [""],
+          "organization": "",
+          "summary": "",
+          "urls": [""]
+        }
+      ],
+      "cve": "",
+      "cwe": {
+        "id": "",
+        "name": ""
+      },
+      "discovery_date": "",
+      "flags": [
+        {
+          "date": "",
+          "group_ids": [""],
+          "label": "",
+          "product_ids": [""]
+        }
+      ],
+      "ids": [
+        {
+          "system_name": "",
+          "text": ""
+        }
+      ],
+      "involvements": [
+        {
+          "date": "",
+          "party": "",
+          "status": "",
+          "summary": ""
+        }
+      ],
+      "notes": [
+        {
+          "audience": "",
+          "category": "",
+          "text": "",
+          "title": ""
+        }
+      ],
+      "product_status": {
+        "first_affected": [""],
+        "first_fixed": [""],
+        "fixed": [""],
+        "known_affected": [""],
+        "known_not_affected": [""],
+        "last_affected": [""],
+        "recommended": [""],
+        "under_investigation": [""]
+      },
+      "references": [
+        {
+          "category": "external",
+          "summary": "",
+          "url": ""
+        }
+      ],
+      "release_date": "",
+      "remediations": [
+        {
+          "category": "",
+          "date": "",
+          "details": "",
+          "entitlements": [""],
+          "group_ids": [""],
+          "product_ids": [""],
+          "restart_required": {
+            "category": "",
+            "details": ""
+          },
+          "url": ""
+        }
+      ],
+      "scores": [
+        {
+          "cvss_v2": {
+            "accessComplexity": "",
+            "accessVector": "",
+            "authentication": "",
+            "availabilityImpact": "",
+            "availabilityRequirement": "",
+            "baseScore": "",
+            "collateralDamagePotential": "",
+            "confidentialityImpact": "",
+            "confidentialityRequirement": "",
+            "environmentalScore": "",
+            "exploitability": "",
+            "integrityImpact": "",
+            "integrityRequirement": "",
+            "remediationLevel": "",
+            "reportConfidence": "",
+            "targetDistribution": "",
+            "temporalScore": "",
+            "vectorString": "",
+            "version": ""
+          },
+          "cvss_v3": {
+            "attackComplexity": "",
+            "attackVector": "",
+            "availabilityImpact": "",
+            "availabilityRequirement": "",
+            "baseScore": "",
+            "baseSeverity": "",
+            "confidentialityImpact": "",
+            "confidentialityRequirement": "",
+            "environmentalScore": "",
+            "environmentalSeverity": "",
+            "exploitCodeMaturity": "",
+            "integrityImpact": "",
+            "integrityRequirement": "",
+            "modifiedAttackComplexity": "",
+            "modifiedAttackVector": "",
+            "modifiedAvailabilityImpact": "",
+            "modifiedConfidentialityImpact": "",
+            "modifiedIntegrityImpact": "",
+            "modifiedPrivilegesRequired": "",
+            "modifiedScope": "",
+            "modifiedUserInteraction": "",
+            "privilegesRequired": "",
+            "remediationLevel": "",
+            "reportConfidence": "",
+            "scope": "",
+            "temporalScore": "",
+            "temporalSeverity": "",
+            "userInteraction": "",
+            "vectorString": "",
+            "version": ""
+          },
+          "products": [""]
+        }
+      ],
+      "threats": [
+        {
+          "category": "",
+          "date": "",
+          "details": "",
+          "group_ids": [""],
+          "product_ids": [""]
+        }
+      ],
+      "title": ""
+    }
+  ]
+}
