+++ app/lib/core/v2_0/doc-min.json
@@ -1,0 +1,6 @@
+{
+  "document": {
+    "category": "csaf_security_advisory",
+    "csaf_version": "2.0"
+  }
+}
